bot/requirements.txt
python-telegram-bot==21.10
python-dotenv==1.1.1
fastapi
uvicorn
